Moved banner logic to its own view model
