+++ WordPress/src/main/java/org/wordpress/android/ui/main/MeFragment.kt
@@ -63,6 +63,7 @@ import org.wordpress.android.ui.deeplinks.DeepLinkOpenWebLinksWithJetpackHelper
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalOverlayUtil
 import org.wordpress.android.ui.main.MeViewModel.RecommendAppUiState
 import org.wordpress.android.ui.main.WPMainActivity.OnScrollToTopListener
+import org.wordpress.android.ui.main.emailverificationbanner.EmailVerificationViewModel
 import org.wordpress.android.ui.main.utils.MeGravatarLoader
 import org.wordpress.android.ui.mysite.jetpackbadge.JetpackPoweredBottomSheetFragment
 import org.wordpress.android.ui.notifications.utils.NotificationsUtils
@@ -162,6 +163,7 @@ class MeFragment : Fragment(R.layout.me_fragment), OnScrollToTopListener {
     lateinit var avatarService: AvatarService
 
     private val viewModel: MeViewModel by viewModels()
+    private val emailVerificationViewModel: EmailVerificationViewModel by viewModels()
 
     private val shouldShowDomainButton
         get() = BuildConfig.IS_JETPACK_APP && domainManagementFeatureConfig.isEnabled() && accountStore.hasAccessToken()
@@ -338,13 +340,13 @@ class MeFragment : Fragment(R.layout.me_fragment), OnScrollToTopListener {
     }
 
     private fun MeFragmentBinding.refreshEmailVerificationBanner(
-        verificationState: MeViewModel.EmailVerificationState
+        verificationState: EmailVerificationViewModel.EmailVerificationState
     ) {
         meEmailVerificationBanner.setVerificationState(
             verificationState,
             accountStore.account.email,
             onLinkClick = {
-                viewModel.onSendVerificationLinkClick(requireActivity())
+                emailVerificationViewModel.onSendVerificationLinkClick(requireActivity())
             }
         )
     }
@@ -389,7 +391,7 @@ class MeFragment : Fragment(R.layout.me_fragment), OnScrollToTopListener {
                 .show(childFragmentManager, JetpackPoweredBottomSheetFragment.TAG)
         }
 
-        viewModel.emailVerificationState.observe(viewLifecycleOwner) { verificationState ->
+        emailVerificationViewModel.emailVerificationState.observe(viewLifecycleOwner) { verificationState ->
             if (!isAdded) return@observe
             refreshEmailVerificationBanner(verificationState)
         }
