+++ WordPress/src/main/java/org/wordpress/android/ui/main/MeViewModel.kt
@@ -1,7 +1,6 @@
 package org.wordpress.android.ui.main
 
 import android.annotation.SuppressLint
-import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
@@ -9,16 +8,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
-import org.greenrobot.eventbus.Subscribe
-import org.greenrobot.eventbus.ThreadMode
 import org.wordpress.android.BuildConfig
 import org.wordpress.android.WordPress
-import org.wordpress.android.fluxc.Dispatcher
-import org.wordpress.android.fluxc.action.AccountAction
-import org.wordpress.android.fluxc.generated.AccountActionBuilder
-import org.wordpress.android.fluxc.store.AccountStore
-import org.wordpress.android.fluxc.store.AccountStore.AccountErrorType
-import org.wordpress.android.fluxc.store.AccountStore.OnAccountChanged
 import org.wordpress.android.models.recommend.RecommendApiCallsProvider
 import org.wordpress.android.models.recommend.RecommendApiCallsProvider.RecommendAppName
 import org.wordpress.android.models.recommend.RecommendApiCallsProvider.RecommendCallResult
@@ -30,7 +21,6 @@ import org.wordpress.android.ui.mysite.SelectedSiteRepository
 import org.wordpress.android.ui.recommend.RecommendAppState
 import org.wordpress.android.ui.recommend.RecommendAppState.ApiFetchedResult
 import org.wordpress.android.ui.recommend.RecommendAppState.FetchingApi
-import org.wordpress.android.util.NetworkUtils
 import org.wordpress.android.util.analytics.AnalyticsUtils.RecommendAppSource.ME
 import org.wordpress.android.util.analytics.AnalyticsUtilsWrapper
 import org.wordpress.android.viewmodel.Event
@@ -46,8 +36,6 @@ class MeViewModel
     private val selectedSiteRepository: SelectedSiteRepository,
     private val recommendApiCallsProvider: RecommendApiCallsProvider,
     private val analyticsUtilsWrapper: AnalyticsUtilsWrapper,
-    private val dispatcher: Dispatcher,
-    private val accountStore: AccountStore,
 ) : ScopedViewModel(mainDispatcher) {
     private val _showDisconnectDialog = MutableLiveData<Event<Boolean>>()
     val showDisconnectDialog: LiveData<Event<Boolean>> = _showDisconnectDialog
@@ -63,13 +51,6 @@ class MeViewModel
 
     private val _showJetpackPoweredBottomSheet = MutableLiveData<Event<Boolean>>()
     val showJetpackPoweredBottomSheet: LiveData<Event<Boolean>> = _showJetpackPoweredBottomSheet
-
-    private val _emailVerificationState = MutableLiveData<EmailVerificationState>()
-    val emailVerificationState: LiveData<EmailVerificationState> = _emailVerificationState
-
-    private var isEmailVerificationLinkRequested: Boolean = false
-    private var isEmailVerificationLinkSent: Boolean = false
-    private var isEmailVerificationError: Boolean = false
 
     data class RecommendAppUiState(
         val showLoading: Boolean = false,
@@ -90,20 +71,6 @@ class MeViewModel
         )
 
         fun isError() = error != null
-    }
-
-    enum class EmailVerificationState {
-        NO_ACCOUNT,     // user doesn't have an account so verification is not possible
-        UNVERIFIED,     // user has not verified their email
-        LINK_REQUESTED, // user has requested a verification link (API call in progress)
-        LINK_SENT,      // verification link has been sent successfully (API call completed)
-        LINK_ERROR,     // an error occurred requesting the verification link
-        VERIFIED,       // user has verified their email address
-    }
-
-    init {
-        dispatcher.register(this)
-        updateEmailVerificationState()
     }
 
     fun signOutWordPress(application: WordPress) {
@@ -153,70 +120,6 @@ class MeViewModel
                 getRecommendTemplate()
             }
         }
-    }
-
-    /*
-     * Update the email verification state, called when we suspect the state has changed
-     */
-    private fun updateEmailVerificationState() {
-        val hasEmail = accountStore.account.email.isNotEmpty()
-        val isEmailVerified = hasEmail && accountStore.account.emailVerified
-        _emailVerificationState.value = if (!isEmailVerified) { // TODO remove the !
-            EmailVerificationState.VERIFIED
-        } else if (isEmailVerificationError) {
-            EmailVerificationState.LINK_ERROR
-        } else if (isEmailVerificationLinkSent) {
-            EmailVerificationState.LINK_SENT
-        } else if (isEmailVerificationLinkRequested) {
-            EmailVerificationState.LINK_REQUESTED
-        } else if (hasEmail) {
-            EmailVerificationState.UNVERIFIED
-        } else {
-            EmailVerificationState.NO_ACCOUNT
-        }
-    }
-
-    /**
-     * User clicked the "Send verification link" button on the email verification banner
-     */
-    fun onSendVerificationLinkClick(context: Context) {
-        if (!NetworkUtils.checkConnection(context)) {
-            return
-        }
-        if (accountStore.hasAccessToken()) {
-            isEmailVerificationError = false
-            isEmailVerificationLinkRequested = true
-            updateEmailVerificationState()
-            // briefly delay the request so the user can see the updated banner if the request completes quickly
-            launch {
-                withContext(bgDispatcher) {
-                    delay(REQUEST_VERIFICATION_LINK_DELAY)
-                    dispatcher.dispatch(AccountActionBuilder.newSendVerificationEmailAction())
-                }
-            }
-        }
-    }
-
-    /**
-     * FluxC event for when the account state changes. Note that we only care about the email verification link
-     * request since that's the only account event sent from the Me screen.
-     */
-    @Suppress("unused")
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun onAccountChanged(event: OnAccountChanged) {
-        if (event.isError) {
-            if (event.error.type == AccountErrorType.SEND_VERIFICATION_EMAIL_ERROR) {
-                isEmailVerificationLinkRequested = false
-                isEmailVerificationLinkSent = false
-                // TODO we ignore event.error because it's blank
-                isEmailVerificationError = true
-            }
-        } else if (event.causeOfChange == AccountAction.SENT_VERIFICATION_EMAIL) {
-            isEmailVerificationLinkSent = true
-            isEmailVerificationError = false
-        }
-
-        updateEmailVerificationState()
     }
 
     private fun getRecommendTemplate() {
@@ -269,6 +172,5 @@ class MeViewModel
 
     companion object {
         private const val SHOW_LOADING_DELAY = 300L
-        private const val REQUEST_VERIFICATION_LINK_DELAY = 750L
     }
 }
